requirements.txt
python-telegram-bot==20.6
spacy==3.7.4
transformers==4.38.2
torch==2.2.1